Move FileHandler to root
